add link to env page
